Add hashmap.py
hashmap.py is an example program which creates a hashmap class.
The hashmap stores key value pairs that can be retrieved in the
same fashion as python dictionaries.
Hashmap contains:
an Array (to store the KV pairs)
a Hash Function (to convert key into an array index)
Collision Handling (append to bucket if hash collision)